// src/types/gae-termination.types.ts
export interface GAETerminationPartial {
  id: number;
  documentId: string;
  terminatedUnits: string;
  terminatedAmount: string;
  terminatedBookingNote: string;
  terminatedCurrency: string;
  terminatedUsdrate: string;
  goldpriceTermination: string;
  approvedByTermination: string;
  terminatedPnl: string;
  statusOfTermination: string;
  terminatedTxn: string | null;
  terminatedTransType: string;
  terminatedCertificate: string;
  terminationReqDate: string;
  managementFeeTerminate: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  locale: string | null;
  paidAmountTerminate: string | null;
  partialToValue: string | null;
  partialFromValue: string | null;
  transaction: {
    id: number;
    documentId: string;
    trDate: string;
    transactionCode: string;
    txnID: string;
    transactionType: string;
    transactionStatus: string;
    rejectReason: string | null;
    fromCurrency: string;
    fromValue: string;
    toCurrency: string;
    toValue: string;
    usdRate: string;
    transactionFee: string;
    createdAt: string;
    updatedAt: string;
    publishedAt: string;
    locale: string | null;
    goldPrice: string;
    paymentMethodUser: string;
    receiptImageLinkUser: string;
    referenceNumberUser: string | null;
    depositDateUser: string | null;
    paymentMethodAdmin: string | null;
    amountAdmin: string | null;
    depositDateAdmin: string | null;
    receiptImageLinkAdmin: string | null;
    referenceNumberAdmin: string | null;
    totalAmountPay: string | null;
    conversionID: string;
    bookingNote: string;
    contractStartDate: string;
    balanceAdmin: string | null;
    amountReceivedAdmin: string | null;
    goldCertificateLink: string | null;
    paymentVoucherLink: string | null;
    managementFeeTerminate: string | null;
    managementFeeAdvance: string;
    managementFeeRate: string | null;
    unitGaeUser: number;
    gaeToBePaid: string;
    gaeDownPayment: string;
    gaeTotal: string;
    contractEndDate: string | null;
    depositedAmount: string | null;
    gaeTotalUnitValue: string;
    goldPriceLME: string;
    isPaid: boolean;
    datePaid: string | null;
    adminWhoPaid: string | null;
    payID: string | null;
    rejectReasonOthers: string | null;
    goldPriceTermination: string | null;
    requestDateForTermination: string | null;
    approvedByTermination: string | null;
    statusOfTermination: string | null;
    approvedByTerminationDate: string | null;
    isRemit: boolean;
    terminatedPnl: string | null;
    terminatedTxn: string | null;
    terminatedAmount: string | null;
    terminatedCurrency: string | null;
    terminateCertificate: string | null;
    receiptImageLinkUserList: string | null;
    transferQmbHash: string | null;
    terminatedUnit: string;
    remainingToValue: string | null;
    remainingBookingNote: string | null;
    remainingFromValue: string | null;
    narrative: string | null;
    remarks: string | null;
    remarkStatus: string | null;
    depositAmount: string | null;
    refund: string | null;
    trackingNumber: string | null;
    isDonate: string | null;
    customer: {
      id: number;
      documentId: string;
      username: string;
      email: string;
      provider: string;
      password: string;
      resetPasswordToken: string | null;
      confirmationToken: string | null;
      confirmed: boolean;
      blocked: boolean;
      country: string;
      phoneNumber: string;
      pin: string;
      kycVerified: string;
      userHash: string;
      createdAt: string;
      updatedAt: string;
      publishedAt: string;
      locale: string | null;
      firstName: string;
      middleName: string | null;
      lastName: string;
      refCode: string;
      isAcceptTerm: boolean;
      resetToken: string | null;
      resetTokenExpiresAt: string | null;
      isVerifyEmail: boolean;
      verifiedCode: string | null;
      suffixName: string;
      is_password_updated: boolean;
      isCharity: string | null;
    };
  };
}